requirements.txt
blinker==1.8.2
cachecontrol==0.14.0
cachetools==5.4.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.0
firebase-admin==6.5.0
flask==3.0.3
google-api-core==2.19.1
google-api-python-client==2.139.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.17.0
google-cloud-storage==2.18.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
grpcio==1.65.2
grpcio-status==1.65.2
httplib2==0.22.0
idna==3.7
importlib-metadata==8.2.0
itsdangerous==2.2.0
jinja2==3.1.4
MarkupSafe==2.1.5
msgpack==1.0.8
proto-plus==1.24.0
protobuf==5.27.2
pyasn1==0.6.0
pyasn1-modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyparsing==3.1.2
requests==2.32.3
rsa==4.9
uritemplate==4.1.1
urllib3==2.2.2
werkzeug==3.0.3
zipp==3.19.2
